// packages/shared-logic/src/logger/types.ts
/**
 * Log level.
 */
export type LogLevel =
  | 'off' // 0
  | 'fatal' // 1
  | 'error' // 2
  | 'success' // 3
  | 'info' // 4
  | 'warn' // 5
  | 'todo' // 6
  | 'debug' // 7
  | 'trace' // 8

/**
 * Log method.
 */
export type LogMethod = 'debug' | 'error' | 'info' | 'log' | 'trace' | 'warn'

/**
 * Entry for log storage.
 */
export interface LogEntry {
  /**
   * Origin.
   */
  src: string

  /**
   * Date.
   */
  date: string

  /**
   * Log level.
   */
  level: LogLevel

  /**
   * Message.
   */
  message: string

  /**
   * Additional information.
   */
  meta?: string

  // Indexer
  [key: string]: string | undefined
}

/**
 * Type of storage.
 */
export type LogStorageType = 'local' | 'remote'

/**
 * Storage function.
 */
export type LogStorageFn = (entry: LogEntry) => void | Promise<void>

/**
 * Log storage.
 */
export interface LogStorage {
  /**
   * Type of storage.
   */
  type: LogStorageType

  /**
   * Max log level to perform operation.
   * Levels that are higher than this level are suppressed.
   */
  level: LogLevel

  /**
   * Storage implementation.
   */
  fn: LogStorageFn
}

/**
 * Options to create a new Logger.
 */
export interface LoggerOptions {
  /**
   * Origin.
   * (Default: "common").
   */
  src?: string

  /**
   * Max log level to print logs to the console.
   * Levels that are higher than this level are suppressed.
   * (Default: "debug" in development, "warn" in production).
   */
  level?: LogLevel

  /**
   * Storages.
   * (Default: []).
   */
  storages?: LogStorage[]
}

/**
 * Options to create a new Logger.
 */
export interface CreateLoggerOptions extends LoggerOptions {
  /**
   * Max log level for local storage.
   */
  local?: LogLevel

  /**
   * Max log level for remote storage.
   */
  remote?: LogLevel
}

/**
 * Log meta.
 */
export type LogMeta = Record<string, unknown> | Error